Replace icons.
