+++ Source/MQTTnet.App/Pages/PacketInspector/PacketView.axaml
@@ -2,9 +2,11 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:packetInspector="clr-namespace:MQTTnet.App.Pages.PacketInspector"
              mc:Ignorable="d"
              d:DesignWidth="800"
              d:DesignHeight="450"
+             d:DataContext="{d:DesignInstance packetInspector:PacketViewModel}"
              x:Class="MQTTnet.App.Pages.PacketInspector.PacketView">
 
     <Grid ColumnDefinitions="Auto,Auto,*,Auto">
@@ -12,6 +14,7 @@
         <Border Background="Gray"
                 Grid.Column="0"
                 Margin="0,0,10,0">
+            
             <TextBlock MinWidth="50"
                        VerticalAlignment="Center"
                        TextAlignment="Right"
@@ -22,12 +25,12 @@
         </Border>
 
         <Image Width="16"
-               Source="/Assets/Icons/arrow_down.png"
+               Source="/Assets/Icons/Input_16x.png"
                Grid.Column="1"
                IsVisible="{Binding IsInbound}" />
 
         <Image Width="16"
-               Source="/Assets/Icons/arrow_up.png"
+               Source="/Assets/Icons/Output_16x.png"
                Grid.Column="1"
                IsVisible="{Binding !IsInbound}" />
 
@@ -42,7 +45,6 @@
             <TextBlock Margin="5,2"
                        Text="5 bytes" />
         </Border>
-
-
+        
     </Grid>
 </UserControl>
